requirements.txt
pytest~=8.3.5
playwright~=1.52.0
pytest-playwright
pytest-bdd~=8.1.0
requests~=2.32.3
jsonschema~=4.23.0
PyYAML~=6.0.2
box~=0.1.5
python-box~=7.3.2
allure-python-commons~=2.14.2